Ajout d'une sécurité pour éviter qu'une personne clique plusieurs fois par accident et que plusieurs requêtes soient envoyées vers le serveur

+++ assets/components/myWalls/Card.js
@@ -1,4 +1,4 @@
-import React, { useState, unmountComponentAtNode } from 'react';
+import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import CardWrapper from './CardWrapper';
 import CardContent from './CardContent';
@@ -10,36 +10,49 @@ import { toast } from 'react-toastify';
 
 function Card({href, title, description, id})
 {
+    //Défini si le menu pour supprimer la carte est ouverte ou non
     const [isDeleteOpen, setIsDeleteOpen] = useState(false);
+    //Défini si la carte est visible ou non (utilisé pour supprimer la carte)
     const [isCardVisible, setIsCardVisible] = useState(true);
-
-    function handleWallDeletion() {
-        //TODO: CSRF
-        axios.delete('/api/my-wall/delete/'+id)
-        .then(function(response){
-            if(response.status === 200)
-            {
-                setIsCardVisible(false);
-                toast.success("Wall removed successfully!")
-            }
-            else
-            {
-                throw new Error;
-            }
-        })
-        .catch(function(error){
-            toast.error('An error occured while trying to delete the wall.');
-        })
+    //Permet de bloquer à 1 click les requêtes vers le serveur lorsque la personne clique sur le bouton de suppression rouge.
+    const [isDeleteRequestProcessed, setIsDeleteRequestProcessed] = useState(false);
 
 
-
+    function handleWallDeletion() 
+    {
+        //On regarde si une requête est déjà en cours pour éviter que la personne clique plusieurs fois par accident et que la personne reçoive 1 message success et plusieurs messages d'erreur par la suite.
+        if(!isDeleteRequestProcessed)
+        {
+            setIsDeleteRequestProcessed(true);
+            
+            //TODO: CSRF
+            axios.delete('/api/my-wall/delete/'+id)
+            .then(function(response){
+                if(response.status === 200)
+                {
+                    setIsCardVisible(false);
+                    toast.success("Wall removed successfully!")
+                }
+                else
+                {
+                    throw new Error;
+                }
+            })
+            .catch(function(error){
+                toast.error('An error occured while trying to delete the wall.');
+            })
+            .finally(function(){
+                setIsDeleteRequestProcessed(false);
+            });
+        }
+        
     }
+
+
 
     function handleDeleteOpen()
     {
         setIsDeleteOpen(!isDeleteOpen);
-        console.log(isDeleteOpen);
-        //todo: cliquer sur le bouton doit pouvoir ouvrir le menu de suppression
     }
 
     return(
